==> examples/1.1_hp_model-exploration.ipynb <==
# Exploratory data analysis

from top2vec import Top2Vec

model = Top2Vec.load("../models/pq-model")

model.get_num_topics()

## Stats

topic_sizes, topic_nums = model.get_topic_sizes()

topic_sizes

topic_nums

---

topic_words, word_scores, topic_nums = model.get_topics(5)

topic_nums

topic_words[0]

word_scores[0]

## Search Topics

keywords = ["ukraine", "russia"]

topic_words, word_scores, topic_scores, topic_nums = model.search_topics(
    keywords=keywords, 
    keywords_neg=[],
    num_topics=5
)

topic_nums

topic_scores

print(f"Topic {topic_nums[0]} was the most similar topic to {keywords} with a cosine similarity of {topic_scores[0]}")
print("Values can be from least similar 0, to most similar 1")

## Search Papers by Topic

documents, document_scores, document_ids = model.search_documents_by_topic(topic_num=126, num_docs=5)

document_ids

document_scores

documents[0][:1_000]

## Search Papers by Keywords

documents, document_scores, document_nums = model.search_documents_by_keywords(
    keywords=keywords, 
#     keywords_neg=[],
    num_docs=20
)

document_nums

documents[0][:1_000]

## Find Similar Words

words, word_scores = model.similar_words(
    keywords=keywords, 
    keywords_neg=[], 
    num_words=20
)

words

word_scores
