==> examples/2.0_hp_topic-reduction.ipynb <==
## Topic-reduction

The model initially found 229 topics which is too much to interpret - thus, we decided to reduce the number of topics using top2vec's hierarchical topic reduction method.

This notebook contains data-preparation part of topic reduction -- for vizzes please go visit this [repository](https://github.com/hcss-utils/streamlit-topic-reduction)

import umap
import pandas as pd
from top2vec import Top2Vec

def reduce_topics(model, min_topics, max_topics, n_words):
    for num in range(min_topics, max_topics+1):
        model.hierarchical_topic_reduction(num)
        topic_words, word_scores, topic_nums = model.get_topics(reduced=True)
        words = [", ".join(topic_words[topic][:n_words]) for topic in topic_nums]
        mapped = [words[num_label] for num_label in model.doc_top_reduced]
        yield {
            num: {
                "unique_words": words,
                "classification_num": model.doc_top_reduced,
                "classification_score": model.doc_dist_reduced,
                "classification_label": mapped,
            }
        }

model = Top2Vec.load("../models/pq-model")

classifications = reduce_topics(model, min_topics=2, max_topics=20, n_words=5)
mappings = {k:v for d in classifications for k, v in d.items()}

umap_model = umap.UMAP(
    n_neighbors=15,
    n_components=2,
    metric="cosine",
    random_state=42
).fit(model._get_document_vectors(norm=False))

embeddings = pd.DataFrame(umap_model.embedding_, columns=("x", "y"))

for num, values in mappings.items():
    embeddings[f"num_label_{num}"] = mappings.get(num)["classification_num"]
    embeddings[f"word_label_{num}"] = mappings.get(num)["classification_label"]
    embeddings[f"score_{num}"] = mappings.get(num)["classification_score"]

embeddings

embeddings.to_csv("../data/processed/deciding-on-topic-reduction.csv", index=False)
