==> examples/2.1_hp_topic-reduction-labeling.ipynb <==
## Labeling

This notebook contains code that export data for manual labeling

import pandas as pd

metadata = pd.read_json("../data/raw/210119_en_deter_preprocessed.json")
metadata = metadata.loc[~metadata["fulltext"].str.contains("^Not available.")].reset_index(drop=True)
embeddings = pd.read_csv("../data/processed/deciding-on-topic-reduction.csv")

assert metadata.shape[0] == embeddings.shape[0]

redundant_cols = ["doc_id", "link", "onestring", "threshold", "pub_title", "lang"]
df = pd.merge(
    metadata.drop(redundant_cols, axis=1),
    embeddings,
    how="left",
    left_index=True,
    right_index=True
)

df.head()

df.to_csv("../data/processed/metadata_embeddings.csv", index=False)

top20 = df.loc[:, ["x", "y", "title", "author", "date", "fulltext", "num_label_20", "word_label_20", "score_20"]]

top20.to_csv("../data/processed/metadata_embeddings_20-topics.csv", index=False)

top_docs = top20.groupby("num_label_20")["score_20"].nlargest(25)

top20.loc[top_docs.index.get_level_values(1), :].to_csv("../data/processed/labeling-20-topics-25-docs.csv", index=False)

---

LABELS = {
    0: "Unidentifiable",
    1: "North Korea nuclear issues",
    2: "India-Pakistani nuclear tensions I", #!
    3: "Middle East nuclear issues",
    4: "Russia's nuclear program",
    5: "Deterrence in the context of the size and reduction of WMD",
    6: "US/NATO rivalry with Russia",
    7: "American internal discourse on deterrence",
    8: "West Europe nuclear issues in the late Cold War context",
    9: "North Korea nuclear issues II", #!
    10: "Nuclear disarmament and non-proliferation",
    11: "Role of the US Air Force in deterrence I", #!
    12: "India-Pakistani nuclear tensions II",
    13: "Cyberdeterrence",
    14: "UK nuclear issues",
    15: "U.S. MPs involvement in defense policy",
    16: "Antiballistic/missile system and deterrence",
    17: "Private companies in nuclear policy",
    18: "Deterrence of China",
    19: "Role of the US Air Force in deterrence II" #!
}

top20["labels"] = top20["num_label_20"].map(LABELS)

top20["tweaked_labels"] = top20["labels"].str.replace("I$|II$", "", regex=True).str.strip()

top20["tweaked_labels"].value_counts()

top20.drop("fulltext", 1).to_csv("../data/processed/labeled-dataset.csv", index=False)
